merged: changes from trunk. revs: r28851:28950
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing2@29097 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/VideoRenderers/LinuxRendererGL.cpp
@@ -948,7 +948,7 @@ void CLinuxRendererGL::UpdateVideoFilter()
   if (m_scalingMethod == VS_SCALINGMETHOD_AUTO)
   {
     bool scaleSD = m_sourceHeight < 720 && m_sourceWidth < 1280;
-    bool scaleUp = m_sourceHeight < g_graphicsContext.GetHeight() && m_sourceWidth < g_graphicsContext.GetWidth();
+    bool scaleUp = (int)m_sourceHeight < g_graphicsContext.GetHeight() && (int)m_sourceWidth < g_graphicsContext.GetWidth();
 
     if (Supports(VS_SCALINGMETHOD_LANCZOS3_FAST) && scaleSD && scaleUp)
       m_scalingMethod = VS_SCALINGMETHOD_LANCZOS3_FAST;
